+++ Panther/src/Context.cpp
@@ -981,25 +981,7 @@ namespace pcit::panther{
 			.allowedInCompile   = true, .allowedInScript   = true, .allowedInBuildSystem = true,
 		};
 
-		this->template_intrinsic_infos[size_t(evo::to_underlying(TemplateIntrinsicFunc::Kind::UI_TO_F))] = 
-		TemplateIntrinsicFuncInfo{
-			.templateParams = evo::SmallVector<TemplateParam>{std::nullopt, std::nullopt},
-			.params         = evo::SmallVector<Param>{Param(AST::FuncDecl::Param::Kind::READ, 0ul)},
-			.returns        = evo::SmallVector<Return>{1ul},
-			.allowedInConstexpr = true, .allowedInComptime = true, .allowedInRuntime     = true,
-			.allowedInCompile   = true, .allowedInScript   = true, .allowedInBuildSystem = true,
-		};
-
 		this->template_intrinsic_infos[size_t(evo::to_underlying(TemplateIntrinsicFunc::Kind::F_TO_I))] = 
-		TemplateIntrinsicFuncInfo{
-			.templateParams = evo::SmallVector<TemplateParam>{std::nullopt, std::nullopt},
-			.params         = evo::SmallVector<Param>{Param(AST::FuncDecl::Param::Kind::READ, 0ul)},
-			.returns        = evo::SmallVector<Return>{1ul},
-			.allowedInConstexpr = true, .allowedInComptime = true, .allowedInRuntime     = true,
-			.allowedInCompile   = true, .allowedInScript   = true, .allowedInBuildSystem = true,
-		};
-
-		this->template_intrinsic_infos[size_t(evo::to_underlying(TemplateIntrinsicFunc::Kind::F_TO_UI))] = 
 		TemplateIntrinsicFuncInfo{
 			.templateParams = evo::SmallVector<TemplateParam>{std::nullopt, std::nullopt},
 			.params         = evo::SmallVector<Param>{Param(AST::FuncDecl::Param::Kind::READ, 0ul)},